libcgroup: Fix compile warning in cgrulesengd.c
We were getting a warning while compiling on 64 bit platforms,

[root@llm72 trunk]# make
cc -std=gnu99 -DDEBUG -g -O2 -I. -DPACKAGE_VERSION=0.32 -Wall -o cgrulesengd cgrulesengd.c \
                -L . -lcgroup -lpthread
cgrulesengd.c: In function ‘cgre_create_netlink_socket_process_msg’:
cgrulesengd.c:327: warning: format ‘%d’ expects type ‘int’, but argument 3 has type ‘long unsigned int’

According to Dan Smith, its just an overzealous format check and so
we just cast the size to integer and print it.

Cc: Dan Smith <danms@us.ibm.com>
Cc: Steve Olivieri <steve86@wpi.edu>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@301 4f4bb910-9a46-0410-90c8-c897d4f1cd53

+++ cgrulesengd.c
@@ -324,7 +324,7 @@ int cgre_create_netlink_socket_process_msg()
 	cn_hdr->ack = 0;
 	cn_hdr->len = sizeof(enum proc_cn_mcast_op);
 	printf("sending netlink message len=%d, cn_msg len=%d\n",
-		nl_hdr->nlmsg_len, sizeof(struct cn_msg));
+		nl_hdr->nlmsg_len, (int) sizeof(struct cn_msg));
 	if (send(sk_nl, nl_hdr, nl_hdr->nlmsg_len, 0) != nl_hdr->nlmsg_len) {
 		printf("failed to send proc connector mcast ctl op!\n");
 		goto close_and_exit;
